index

+++ src/YiAim.Cms.Application.Contracts/Blogs/Blog/BaseBlogDto.cs
@@ -8,7 +8,7 @@ public class BaseBlogDto : EntityDto<long>
 {
     public string Title { get; set; }
     public int Taxis { get; set; }
-    public int? CategoryId { get; set; }
+    public long? CategoryId { get; set; }
     public long PublishDate { get; set; }
     public string Author { get; set; }
     public bool IsHot { get; set; }
